+++ package.js
@@ -1,7 +1,7 @@
 Package.describe({
     summary: "Standalone test spies, stubs and mocks for JavaScript.",
     name: "practicalmeteor:sinon",
-    version: "1.10.3_2",
+    version: "1.14.1",
     git: "https://github.com/practicalmeteor/meteor-sinon.git"
 });
 
@@ -11,7 +11,7 @@ Package.onUse(function (api, where) {
     api.use('coffeescript');
     api.use('practicalmeteor:chai@1.9.2_3');
 
-    api.addFiles('sinon-1.10.3.js');
+    api.addFiles('sinon-1.14.1.js');
     api.addFiles('sinon-chai-2.6.0.js');
     api.addFiles('install-sinonChai.js');
     api.addFiles('Helpers.coffee');
